added gamelister, add new feats
Add 'gamelister', which will scan your current directory and print a .txt file of current games. Could be useful for processing tools like AI. Additionally, I am experimenting with tighter accurate file identifications for the ROM mover. I've also added a 'results' folder with logs in it everytime it runs to allow the user to send this to an external processor for further analysis.